reach final interpolation value + smoother

+++ src/engine_camera.h
@@ -86,17 +86,17 @@ struct Camera {
 extern long previous_cam_mappos_x;
 extern long previous_cam_mappos_y;
 extern long previous_cam_mappos_z;
-extern long interpolated_cam_mappos_x;
-extern long interpolated_cam_mappos_y;
-extern long interpolated_cam_mappos_z;
-extern long previous_cam_rotation_angle_x;
-extern long previous_cam_rotation_angle_y;
-extern long previous_cam_rotation_angle_z;
-extern long interpolated_cam_rotation_angle_x;
-extern long interpolated_cam_rotation_angle_y;
-extern long interpolated_cam_rotation_angle_z;
+extern float interpolated_cam_mappos_x;
+extern float interpolated_cam_mappos_y;
+extern float interpolated_cam_mappos_z;
+extern float previous_cam_rotation_angle_x;
+extern float previous_cam_rotation_angle_y;
+extern float previous_cam_rotation_angle_z;
+extern float interpolated_cam_rotation_angle_x;
+extern float interpolated_cam_rotation_angle_y;
+extern float interpolated_cam_rotation_angle_z;
 extern long previous_camera_zoom;
-extern long interpolated_camera_zoom;
+extern float interpolated_camera_zoom;
 
 /******************************************************************************/
